@@ -17,10 +17,10 @@ class PointerThisPlease<T> {
 
 Widget prepareWidget(dynamic object,
     {dynamic parameter = const NotGiven(),
-      BuildContext context,
-      Function stringToWidgetFunction}) {
+      BuildContext? context,
+      Function? stringToWidgetFunction}) {
   if (object == null) {
-    return (null);
+    return (SizedBox.fromSize());
   }
   if (object is Widget) {
     return (object);
@@ -54,9 +54,9 @@ Widget prepareWidget(dynamic object,
 
 class SearchableDropdown<T> extends StatefulWidget {
   final List<DropdownMenuItem<T>> items;
-  final Function onChanged;
-  final T value;
-  final TextStyle style;
+  final Function? onChanged;
+  final T? value;
+  final TextStyle? style;
   final dynamic searchHint;
   final dynamic hint;
   final dynamic disabledHint;
@@ -68,24 +68,24 @@ class SearchableDropdown<T> extends StatefulWidget {
   final bool isShowSelectAllOrNone;
   final bool displayClearIcon;
   final Icon clearIcon;
-  final Color iconEnabledColor;
-  final Color iconDisabledColor;
+  final Color? iconEnabledColor;
+  final Color? iconDisabledColor;
   final double iconSize;
   final bool isExpanded;
   final bool isCommaSeparated;
   final bool isCaseSensitiveSearch;
-  final Function searchFn;
-  final Function onClear;
-  final Function selectedValueWidgetFn;
+  final Function? searchFn;
+  final Function? onClear;
+  final Function? selectedValueWidgetFn;
   final TextInputType keyboardType;
-  final Function validator;
+  final Function? validator;
   final bool multipleSelection;
   final List<int> selectedItems;
-  final Function displayItem;
+  final Function? displayItem;
   final bool dialogBox;
-  final BoxConstraints menuConstraints;
+  final BoxConstraints? menuConstraints;
   final bool readOnly;
-  final Color menuBackgroundColor;
+  final Color? menuBackgroundColor;
 
   /// Search choices Widget with a single choice that opens a dialog or a menu to let the user do the selection conveniently with a search.
   ///
@@ -120,11 +120,11 @@ class SearchableDropdown<T> extends StatefulWidget {
   /// @param readOnly [bool] whether to let the user choose the value to select or just present the selected value if any.
   /// @param menuBackgroundColor [Color] background color of the menu whether in dialog box or menu mode.
   factory SearchableDropdown.single({
-    Key key,
-    @required List<DropdownMenuItem<T>> items,
-    @required Function onChanged,
-    T value,
-    TextStyle style,
+    Key? key,
+    required List<DropdownMenuItem<T>> items,
+    Function? onChanged,
+    T? value,
+    TextStyle? style,
     dynamic searchHint,
     dynamic hint,
     dynamic disabledHint,
@@ -135,22 +135,22 @@ class SearchableDropdown<T> extends StatefulWidget {
     dynamic closeButton = "Close",
     bool displayClearIcon = true,
     Icon clearIcon = const Icon(Icons.clear),
-    Color iconEnabledColor,
-    Color iconDisabledColor,
+    Color? iconEnabledColor,
+    Color? iconDisabledColor,
     double iconSize = 24.0,
     bool isExpanded = false,
     bool isCaseSensitiveSearch = false,
-    Function searchFn,
-    Function onClear,
-    Function selectedValueWidgetFn,
+    Function? searchFn,
+    Function? onClear,
+    Function? selectedValueWidgetFn,
     TextInputType keyboardType = TextInputType.text,
-    Function validator,
+    Function? validator,
     bool assertUniqueValue = true,
-    Function displayItem,
+    Function? displayItem,
     bool dialogBox = true,
-    BoxConstraints menuConstraints,
+    BoxConstraints? menuConstraints,
     bool readOnly = false,
-    Color menuBackgroundColor,
+    Color? menuBackgroundColor,
   }) {
     return (SearchableDropdown._(
       key: key,
@@ -219,11 +219,11 @@ class SearchableDropdown<T> extends StatefulWidget {
   /// @param readOnly [bool] whether to let the user choose the value to select or just present the selected value if any.
   /// @param menuBackgroundColor [Color] background color of the menu whether in dialog box or menu mode.
   factory SearchableDropdown.multiple({
-    Key key,
-    @required List<DropdownMenuItem<T>> items,
-    @required Function onChanged,
+    Key? key,
+    required List<DropdownMenuItem<T>> items,
+    required Function onChanged,
     List<int> selectedItems = const [],
-    TextStyle style,
+    TextStyle? style,
     dynamic searchHint,
     dynamic hint,
     dynamic disabledHint,
@@ -235,22 +235,22 @@ class SearchableDropdown<T> extends StatefulWidget {
     bool isShowSelectAllOrNone = false,
     bool displayClearIcon = true,
     Icon clearIcon = const Icon(Icons.clear),
-    Color iconEnabledColor,
-    Color iconDisabledColor,
+    Color? iconEnabledColor,
+    Color? iconDisabledColor,
     double iconSize = 24.0,
     bool isExpanded = false,
     bool isCommaSeparated = false,
     bool isCaseSensitiveSearch = false,
-    Function searchFn,
-    Function onClear,
-    Function selectedValueWidgetFn,
+    Function? searchFn,
+    Function? onClear,
+    Function? selectedValueWidgetFn,
     TextInputType keyboardType = TextInputType.text,
-    Function validator,
-    Function displayItem,
+    Function? validator,
+    Function? displayItem,
     bool dialogBox = true,
-    BoxConstraints menuConstraints,
+    BoxConstraints? menuConstraints,
     bool readOnly = false,
-    Color menuBackgroundColor,
+    Color? menuBackgroundColor,
   }) {
     return (SearchableDropdown._(
       key: key,
@@ -290,8 +290,8 @@ class SearchableDropdown<T> extends StatefulWidget {
   }
 
   SearchableDropdown._({
-    Key key,
-    @required this.items,
+    Key? key,
+    required this.items,
     this.onChanged,
     this.value,
     this.style,
@@ -320,7 +320,7 @@ class SearchableDropdown<T> extends StatefulWidget {
     this.selectedItems = const [],
     this.doneButton,
     this.displayItem,
-    this.dialogBox,
+    required this.dialogBox,
     this.menuConstraints,
     this.readOnly = false,
     this.menuBackgroundColor,
@@ -332,9 +332,9 @@ class SearchableDropdown<T> extends StatefulWidget {
         super(key: key);
 
   SearchableDropdown({
-    Key key,
-    @required this.items,
-    @required this.onChanged,
+    Key? key,
+    required this.items,
+    this.onChanged,
     this.value,
     this.style,
     this.searchHint,
@@ -378,25 +378,25 @@ class SearchableDropdown<T> extends StatefulWidget {
 }
 
 class _SearchableDropdownState<T> extends State<SearchableDropdown<T>> {
-  List<int> selectedItems;
+  List<int>? selectedItems;
   PointerThisPlease<bool> displayMenu = PointerThisPlease<bool>(false);
 
   TextStyle get _textStyle =>
       widget.style ??
           (_enabled && !(widget.readOnly ?? false)
-              ? Theme.of(context).textTheme.subtitle1
+              ? Theme.of(context).textTheme.titleMedium
               : Theme.of(context)
               .textTheme
-              .subtitle1
-              .copyWith(color: _disabledIconColor));
+              .titleMedium
+              ?.copyWith(color: _disabledIconColor)) ?? TextStyle();
   bool get _enabled =>
       widget.items != null &&
-          widget.items.isNotEmpty &&
+          widget.items!.isNotEmpty &&
           widget.onChanged != null;
 
   Color get _enabledIconColor {
     if (widget.iconEnabledColor != null) {
-      return widget.iconEnabledColor;
+      return widget.iconEnabledColor!;
     }
     switch (Theme.of(context).brightness) {
       case Brightness.light:
@@ -409,7 +409,7 @@ class _SearchableDropdownState<T> extends State<SearchableDropdown<T>> {
 
   Color get _disabledIconColor {
     if (widget.iconDisabledColor != null) {
-      return widget.iconDisabledColor;
+      return widget.iconDisabledColor!;
     }
     switch (Theme.of(context).brightness) {
       case Brightness.light:
@@ -431,23 +431,23 @@ class _SearchableDropdownState<T> extends State<SearchableDropdown<T>> {
     if (widget.validator == null) {
       return (true);
     }
-    return (widget.validator(selectedResult) == null);
+    return (widget.validator!(selectedResult) == null);
   }
 
   bool get hasSelection {
-    return (selectedItems != null && selectedItems.isNotEmpty);
+    return (selectedItems != null && selectedItems!.isNotEmpty);
   }
 
   dynamic get selectedResult {
     return (widget.multipleSelection
         ? selectedItems
         : selectedItems?.isNotEmpty ?? false
-        ? widget.items[selectedItems.first]?.value
+        ? widget.items[selectedItems!.first].value
         : null);
   }
 
-  int indexFromValue(T value) {
-    return (widget.items.indexWhere((item) {
+  int? indexFromValue(T value) {
+    return (widget.items?.indexWhere((item) {
       return (item.value == value);
     }));
   }
@@ -465,7 +465,7 @@ class _SearchableDropdownState<T> extends State<SearchableDropdown<T>> {
     if (widget.multipleSelection) {
       selectedItems = List<int>.from(widget.selectedItems ?? []);
     } else if (widget.value != null) {
-      int i = indexFromValue(widget.value);
+      int? i = indexFromValue(widget.value!);
       if (i != null && i != -1) {
         selectedItems = [i];
       }
@@ -474,7 +474,7 @@ class _SearchableDropdownState<T> extends State<SearchableDropdown<T>> {
   }
 
   @override
-  void didUpdateWidget(SearchableDropdown oldWidget) {
+  void didUpdateWidget(SearchableDropdown<T> oldWidget) {
     super.didUpdateWidget(oldWidget);
     _updateSelectedIndex();
   }
@@ -498,10 +498,10 @@ class _SearchableDropdownState<T> extends State<SearchableDropdown<T>> {
       menuConstraints: widget.menuConstraints,
       menuBackgroundColor: widget.menuBackgroundColor,
       callOnPop: () {
-        if (!widget.dialogBox &&
+        if (widget.dialogBox != null && !widget.dialogBox! &&
             widget.onChanged != null &&
             selectedItems != null) {
-          widget.onChanged(selectedResult);
+          widget.onChanged!(selectedResult);
         }
         setState(() {});
       },
@@ -512,14 +512,14 @@ class _SearchableDropdownState<T> extends State<SearchableDropdown<T>> {
   Widget build(BuildContext context) {
     final List<Widget> items =
     _enabled ? List<Widget>.from(widget.items) : <Widget>[];
-    int hintIndex;
+    int? hintIndex;
     if (widget.hint != null ||
         (!_enabled && prepareWidget(widget.disabledHint) != null)) {
-      final Widget emplacedHint = _enabled
+      final Widget? emplacedHint = _enabled
           ? prepareWidget(widget.hint)
           : DropdownMenuItem<Widget>(
           child: prepareWidget(widget.disabledHint) ??
-              prepareWidget(widget.hint));
+              prepareWidget(widget.hint) ?? SizedBox());
       hintIndex = items.length;
       items.add(DefaultTextStyle(
         style: _textStyle.copyWith(color: Theme.of(context).hintColor),
@@ -530,17 +530,17 @@ class _SearchableDropdownState<T> extends State<SearchableDropdown<T>> {
       ));
     }
     Widget innerItemsWidget;
-    List<Widget> list = List<Widget>();
+    List<Widget> list = <Widget>[];
     if(widget.isCommaSeparated){
       if(widget.selectedValueWidgetFn != null){
-        if (selectedItems!=null&&selectedItems.length>0 ) {
+        if (selectedItems!=null&&selectedItems!.length>0 ) {
           List<DropdownMenuItem<T>> selectdeDropdownitems=[];
-          for (int item in selectedItems) {
+          for (int item in selectedItems!) {
             if(widget.items.length>item){
               selectdeDropdownitems.add(widget.items[item]);
             }
           }
-          list.add(widget.selectedValueWidgetFn(selectdeDropdownitems));
+          list.add(widget.selectedValueWidgetFn!(selectdeDropdownitems));
         }
       }else{
         selectedItems?.forEach((item) {
@@ -550,7 +550,7 @@ class _SearchableDropdownState<T> extends State<SearchableDropdown<T>> {
     }else{
       selectedItems?.forEach((item) {
         list.add(widget.selectedValueWidgetFn != null
-            ? widget.selectedValueWidgetFn(widget.items[item].value)
+            ? widget.selectedValueWidgetFn!(widget.items[item].value)
             : items[item]);
       });
     }
@@ -571,7 +571,7 @@ class _SearchableDropdownState<T> extends State<SearchableDropdown<T>> {
         onTap: (widget.readOnly ?? false) || !_enabled
             ? null
             : () async {
-          if (widget.dialogBox) {
+          if (widget.dialogBox ?? false) {
             await showDialog(
                 context: context,
                 barrierDismissible: true,
@@ -579,12 +579,14 @@ class _SearchableDropdownState<T> extends State<SearchableDropdown<T>> {
                   return (menuWidget);
                 });
             if (widget.onChanged != null && selectedItems != null) {
-              widget.onChanged(selectedResult);
+              widget.onChanged!(selectedResult);
             }
           } else {
             displayMenu.value = true;
           }
-          setState(() {});
+          if (mounted) {
+            setState(() {});
+          }
         },
         child: Row(
           children: <Widget>[
@@ -647,7 +649,7 @@ class _SearchableDropdownState<T> extends State<SearchableDropdown<T>> {
     final double bottom = 8.0;
     var validatorOutput;
     if (widget.validator != null) {
-      validatorOutput = widget.validator(selectedResult);
+      validatorOutput = widget.validator!(selectedResult);
     }
     var labelOutput = prepareWidget(widget.label, parameter: selectedResult,
         stringToWidgetFunction: (string) {
@@ -699,38 +701,38 @@ class _SearchableDropdownState<T> extends State<SearchableDropdown<T>> {
   }
 
   clearSelection() {
-    selectedItems.clear();
+    selectedItems?.clear();
     if (widget.onChanged != null) {
-      widget.onChanged(selectedResult);
+      widget.onChanged!(selectedResult);
     }
     if (widget.onClear != null) {
-      widget.onClear();
+      widget.onClear!();
     }
     setState(() {});
   }
 }
 
 class DropdownDialog<T> extends StatefulWidget {
-  final List<DropdownMenuItem<T>> items;
-  final Widget hint;
+  final List<DropdownMenuItem<T>>? items;
+  final Widget? hint;
   final bool isCaseSensitiveSearch;
   final dynamic closeButton;
   final bool isShowSelectAllOrNone;
-  final TextInputType keyboardType;
-  final Function searchFn;
+  final TextInputType? keyboardType;
+  final Function? searchFn;
   final bool multipleSelection;
-  final List<int> selectedItems;
-  final Function displayItem;
+  final List<int>? selectedItems;
+  final Function? displayItem;
   final dynamic doneButton;
-  final Function validator;
-  final bool dialogBox;
-  final PointerThisPlease<bool> displayMenu;
-  final BoxConstraints menuConstraints;
-  final Function callOnPop;
-  final Color menuBackgroundColor;
+  final Function? validator;
+  final bool? dialogBox;
+  final PointerThisPlease<bool>? displayMenu;
+  final BoxConstraints? menuConstraints;
+  final Function? callOnPop;
+  final Color? menuBackgroundColor;
 
   DropdownDialog({
-    Key key,
+    Key? key,
     this.items,
     this.hint,
     this.isCaseSensitiveSearch = false,
@@ -738,7 +740,7 @@ class DropdownDialog<T> extends StatefulWidget {
     this.isShowSelectAllOrNone =false,
     this.keyboardType,
     this.searchFn,
-    this.multipleSelection,
+    required this.multipleSelection,
     this.selectedItems,
     this.displayItem,
     this.doneButton,
@@ -759,7 +761,7 @@ class _DropdownDialogState<T> extends State<DropdownDialog> {
   TextStyle defaultButtonStyle =
   new TextStyle(fontSize: 16, fontWeight: FontWeight.w500);
   List<int> shownIndexes = [];
-  Function searchFn;
+  late Function searchFn;
 
   _DropdownDialogState();
 
@@ -767,7 +769,7 @@ class _DropdownDialogState<T> extends State<DropdownDialog> {
     return (widget.multipleSelection
         ? widget.selectedItems
         : widget.selectedItems?.isNotEmpty ?? false
-        ? widget.items[widget.selectedItems.first]?.value
+        ? widget.items![widget.selectedItems!.first]?.value
         : null);
   }
 
@@ -778,7 +780,7 @@ class _DropdownDialogState<T> extends State<DropdownDialog> {
   @override
   void initState() {
     if (widget.searchFn != null) {
-      searchFn = widget.searchFn;
+      searchFn = widget.searchFn!;
     } else {
       Function matchFn;
       if (widget.isCaseSensitiveSearch) {
@@ -796,7 +798,7 @@ class _DropdownDialogState<T> extends State<DropdownDialog> {
       searchFn = (keyword, items) {
         List<int> shownIndexes = [];
         int i = 0;
-        widget.items.forEach((item) {
+        widget.items?.forEach((item) {
           if (matchFn(item, keyword) || (keyword?.isEmpty ?? true)) {
             shownIndexes.add(i);
           }
@@ -819,8 +821,8 @@ class _DropdownDialogState<T> extends State<DropdownDialog> {
       child: new Card(
         color: widget.menuBackgroundColor,
         margin: EdgeInsets.symmetric(
-            vertical: widget.dialogBox ? 10 : 5,
-            horizontal: widget.dialogBox ? 10 : 4),
+            vertical: (widget.dialogBox) ?? false ? 10 : 5,
+            horizontal: (widget.dialogBox ?? false) ? 10 : 4),
         child: new Container(
           constraints: widget.menuConstraints,
           padding: EdgeInsets.symmetric(vertical: 15, horizontal: 15),
@@ -844,13 +846,13 @@ class _DropdownDialogState<T> extends State<DropdownDialog> {
     if (widget.validator == null) {
       return (true);
     }
-    return (widget.validator(selectedResult) == null);
+    return (widget.validator!(selectedResult) == null);
   }
 
   Widget titleBar() {
     var validatorOutput;
     if (widget.validator != null) {
-      validatorOutput = widget.validator(selectedResult);
+      validatorOutput = widget.validator!(selectedResult);
     }
 
     Widget validatorOutputWidget = valid
@@ -958,12 +960,12 @@ class _DropdownDialogState<T> extends State<DropdownDialog> {
   }
 
   pop() {
-    if (widget.dialogBox) {
+    if (widget.dialogBox ?? false) {
       Navigator.pop(context);
     } else {
-      widget.displayMenu.value = false;
+      widget.displayMenu?.value = false;
       if (widget.callOnPop != null) {
-        widget.callOnPop();
+        widget.callOnPop!();
       }
     }
   }
@@ -973,20 +975,20 @@ class _DropdownDialogState<T> extends State<DropdownDialog> {
       child: Scrollbar(
         child: new ListView.builder(
           itemBuilder: (context, index) {
-            DropdownMenuItem item = widget.items[shownIndexes[index]];
+            DropdownMenuItem item = widget.items![shownIndexes[index]];
             return new InkWell(
               onTap: () {
                 if (widget.multipleSelection) {
                   setState(() {
-                    if (widget.selectedItems.contains(shownIndexes[index])) {
-                      widget.selectedItems.remove(shownIndexes[index]);
+                    if (widget.selectedItems?.contains(shownIndexes[index]) ?? false) {
+                      widget.selectedItems?.remove(shownIndexes[index]);
                     } else {
-                      widget.selectedItems.add(shownIndexes[index]);
+                      widget.selectedItems?.add(shownIndexes[index]);
                     }
                   });
                 } else {
-                  widget.selectedItems.clear();
-                  widget.selectedItems.add(shownIndexes[index]);
+                  widget.selectedItems?.clear();
+                  widget.selectedItems?.add(shownIndexes[index]);
                   if (widget.doneButton == null) {
                     pop();
                   } else {
@@ -998,7 +1000,7 @@ class _DropdownDialogState<T> extends State<DropdownDialog> {
                   ? widget.displayItem == null
                   ? (Row(children: [
                 Icon(
-                  widget.selectedItems.contains(shownIndexes[index])
+                  (widget.selectedItems?.contains(shownIndexes[index]) ?? false)
                       ? Icons.check_box
                       : Icons.check_box_outline_blank,
                 ),
@@ -1007,11 +1009,11 @@ class _DropdownDialogState<T> extends State<DropdownDialog> {
                 ),
                 Flexible(child: item),
               ]))
-                  : widget.displayItem(item,
-                  widget.selectedItems.contains(shownIndexes[index]))
+                  : widget.displayItem!(item,
+                  widget.selectedItems?.contains(shownIndexes[index]))
                   : widget.displayItem == null
                   ? item
-                  : widget.displayItem(item, item.value == selectedResult),
+                  : widget.displayItem!(item, item.value == selectedResult),
             );
           },
           itemCount: shownIndexes.length,
@@ -1060,7 +1062,7 @@ class _DropdownDialogState<T> extends State<DropdownDialog> {
                     onPressed: () {
                       setState(() {
                         selectedResult.clear();
-                        selectedResult.addAll(Iterable<int>.generate(widget.items.length).toList());
+                        selectedResult.addAll(Iterable<int>.generate(widget.items?.length ?? 0).toList());
                       });
                     },
                     child: Text("Select all")),
